Split HTTP Methods from ActionDispatch::Request to seperate module Add trace? method to created module
